only have constants used in multiple files in consts.py
